# frontend/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CarDian",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/lookups/manual/": {
            "post": {
                "tags": [
                    "Manual Lookups"
                ],
                "summary": "New manual lookup",
                "operationId": "manual_lookup_v1_lookups_manual__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ManualLookupRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ManualLookupAcceptedResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/lookups/manual/{lookup_id}/verdict": {
            "get": {
                "tags": [
                    "Manual Lookups"
                ],
                "summary": "Get manual lookup verdict (polling)",
                "operationId": "get_verdict_manual_v1_lookups_manual__lookup_id__verdict_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "lookup_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Lookup Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerdictResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/lookups/manual/{lookup_id}": {
            "get": {
                "tags": [
                    "Manual Lookups"
                ],
                "summary": "Get lookup by id",
                "operationId": "get_users_manual_lookup_by_id_v1_lookups_manual__lookup_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "lookup_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Lookup Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ManualLookupResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/lookups/manual/my": {
            "get": {
                "tags": [
                    "Manual Lookups"
                ],
                "summary": "Get all users manual lookups",
                "operationId": "get_all_users_manual_lookups_v1_lookups_manual_my_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SequenceManualLookupResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/lookups/parsed/": {
            "post": {
                "tags": [
                    "Parsed Lookups"
                ],
                "summary": "New parsed lookups",
                "operationId": "new_parsed_lookup_v1_lookups_parsed__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParsedLookupsRequestSchema"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParsedLookupAcceptedSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/v1/lookups/parsed/{lookup_id}/verdict": {
            "get": {
                "tags": [
                    "Parsed Lookups"
                ],
                "summary": "Get parsed lookup verdict (polling)",
                "operationId": "get_verdict_parsed_v1_lookups_parsed__lookup_id__verdict_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "lookup_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Lookup Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VerdictResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/lookups/parsed/{lookup_id}": {
            "get": {
                "tags": [
                    "Parsed Lookups"
                ],
                "summary": "Get parsed lookup by id",
                "operationId": "get_users_parsed_lookup_by_id_v1_lookups_parsed__lookup_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "lookup_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Lookup Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ParsedLookupsResponseSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/lookups/parsed/my": {
            "get": {
                "tags": [
                    "Parsed Lookups"
                ],
                "summary": "Get all users parsed lookups",
                "operationId": "get_all_users_parsed_lookups_v1_lookups_parsed_my_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SequenceParsedLookupResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
        "/v1/watchlists/my/all": {
            "get": {
                "tags": [
                    "Watchlists"
                ],
                "summary": "Get user watchlist (including inactive)",
                "operationId": "get_users_watchlist_v1_watchlists_my_all_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchlistSequenceResponseSchema"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ManualLookupAcceptedResponseSchema": {
                "properties": {
                    "lookup_id": {
                        "type": "integer",
                        "title": "Lookup Id"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "task_id": {
                        "type": "string",
                        "title": "Task Id"
                    },
                    "status_code": {
                        "type": "integer",
                        "title": "Status Code"
                    }
                },
                "type": "object",
                "required": [
                    "lookup_id",
                    "user_id",
                    "task_id",
                    "status_code"
                ],
                "title": "ManualLookupAcceptedResponseSchema"
            },
            "ManualLookupRequestSchema": {
                "properties": {
                    "brand": {
                        "type": "string",
                        "title": "Brand"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "year": {
                        "type": "integer",
                        "exclusiveMinimum": 1850.0,
                        "title": "Year"
                    },
                    "mileage_km": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Mileage Km"
                    },
                    "fuel_category": {
                        "$ref": "#/components/schemas/FuelCategories"
                    },
                    "transmission": {
                        "$ref": "#/components/schemas/TransmissionType"
                    },
                    "condition": {
                        "$ref": "#/components/schemas/Condition"
                    },
                    "power_kw": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Power Kw"
                    },
                    "body_type": {
                        "$ref": "#/components/schemas/BodyTypes"
                    },
                    "drive_train": {
                        "$ref": "#/components/schemas/DriveTrainTypes"
                    },
                    "had_accident": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "has_full_service_history": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "previous_owners_qty": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Previous Owners Qty"
                    },
                    "seller_is_dealer": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    }
                },
                "type": "object",
                "required": [
                    "brand",
                    "model",
                    "year",
                    "mileage_km",
                    "fuel_category",
                    "transmission",
                    "condition",
                    "power_kw",
                    "body_type",
                    "drive_train",
                    "had_accident",
                    "has_full_service_history",
                    "previous_owners_qty",
                    "seller_is_dealer",
                    "user_id"
                ],
                "title": "ManualLookupRequestSchema"
            },
            "ManualLookupResponseSchema": {
                "properties": {
                    "brand": {
                        "type": "string",
                        "title": "Brand"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model"
                    },
                    "year": {
                        "type": "integer",
                        "exclusiveMinimum": 1850.0,
                        "title": "Year"
                    },
                    "mileage_km": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Mileage Km"
                    },
                    "fuel_category": {
                        "$ref": "#/components/schemas/FuelCategories"
                    },
                    "transmission": {
                        "$ref": "#/components/schemas/TransmissionType"
                    },
                    "condition": {
                        "$ref": "#/components/schemas/Condition"
                    },
                    "power_kw": {
                        "type": "integer",
                        "exclusiveMinimum": 0.0,
                        "title": "Power Kw"
                    },
                    "body_type": {
                        "$ref": "#/components/schemas/BodyTypes"
                    },
                    "drive_train": {
                        "$ref": "#/components/schemas/DriveTrainTypes"
                    },
                    "had_accident": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "has_full_service_history": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "previous_owners_qty": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Previous Owners Qty"
                    },
                    "seller_is_dealer": {
                        "$ref": "#/components/schemas/BoolType"
                    },
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "price_listed": {
                        "type": "integer",
                        "minimum": 0.0,
                        "title": "Price Listed"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "brand",
                    "model",
                    "year",
                    "mileage_km",
                    "fuel_category",
                    "transmission",
                    "condition",
                    "power_kw",
                    "body_type",
                    "drive_train",
                    "had_accident",
                    "has_full_service_history",
                    "previous_owners_qty",
                    "seller_is_dealer",
                    "user_id",
                    "id",
                    "status",
                    "price_listed",
                    "created_at"
                ],
                "title": "ManualLookupResponseSchema"
            },
...
            "SequenceManualLookupResponseSchema": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ManualLookupResponseSchema"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total_items_qty": {
                        "type": "integer",
                        "title": "Total Items Qty"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total_items_qty"
                ],
                "title": "SequenceManualLookupResponseSchema"
            },
...
            "TransmissionType": {
                "type": "string",
                "enum": [
                    "automatic",
                    "semi_automathic",
                    "manual"
                ],
                "title": "TransmissionType"
            },
...
            "VerdictResponseSchema": {
                "properties": {
                    "predicted_price": {
                        "type": "integer",
                        "title": "Predicted Price"
                    },
                    "price_deviation_percent": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Deviation Percent"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "lower_bar": {
                        "type": "integer",
                        "title": "Lower Bar"
                    },
                    "upper_bar": {
                        "type": "integer",
                        "title": "Upper Bar"
                    },
                    "llm_feedback": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Llm Feedback"
                    },
                    "manual_lookup_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Manual Lookup Id"
                    },
                    "parsed_lookup_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parsed Lookup Id"
                    }
                },
                "type": "object",
                "required": [
                    "predicted_price",
                    "id",
                    "lower_bar",
                    "upper_bar"
                ],
                "title": "VerdictResponseSchema"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/v1/auth/login"
                    }
                }
            }
        }
    }
}